[fix] 예상하지 못했던 예외 / 실제 동작에서 발생하는 에러들 처리

+++ src/main/java/subway/station/StationDao.java
@@ -6,11 +6,17 @@ import org.springframework.stereotype.Repository;
 import subway.exceptions.DuplicateStationException;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import subway.section.SectionDao;
+import subway.exceptions.BadRequestException;
+import org.springframework.beans.factory.annotation.Autowired;
+import subway.exceptions.*;
 
 import java.util.List;
 
 @Repository
 public class StationDao {
+    @Autowired
+    private SectionDao sectionDao;
     private JdbcTemplate jdbcTemplate;
     private SimpleJdbcInsert simpleJdbcInsert;
 
@@ -29,8 +35,8 @@ public class StationDao {
             Number id = simpleJdbcInsert.executeAndReturnKey(params);
 
             return findById(id.longValue());
-        } catch (DataIntegrityViolationException e){
-            throw new DuplicateStationException();
+        } catch (BadRequestException e){
+            throw new BadRequestException();
         }
     }
 
@@ -46,6 +52,9 @@ public class StationDao {
     }
 
     public void deleteById(Long id) {
+        if (sectionDao.findByStationId(id).size() > 0 ){
+            throw new BadRequestException();
+        }
         jdbcTemplate.update("delete from STATION where id = ?", id);
     }
 }
